Move dss.storage.index_document to dss.index.es.document
